- fused success and failure mockup behaviour
