+++ alica_test_utility/src/SuccessOrFailBehaviour.cpp
@@ -1,0 +1,1 @@
+#include "alica/test/SuccessOrFailBehaviour.hpp"
